Добавила модальное окно в хедере
