+++ src/index.html
@@ -178,15 +178,57 @@
       </div>
     </footer>
 
+    <!-- modal header -->
+    <div class="js-lightbox-header modal-lightbox">
+      <div class="lightbox-overlay-header overlay">
+        <div class="lightbox-content-header content-modal">
+          <form class="modal-form">
+            <h2 class="modal-form-title-header">
+              Sing in
+            </h2>
+            <div class="button-check-in">
+              <button class="form-button button-submit">Sing in from Facebook</button>
+              <button class="form-button button-submit">Sing in from Gmail</button>
+            </div>
+            <div class="modal-form-input-wrapper">
+              <label for="user-email-form" class="modal-form-label">Email</label>
+              <input 
+              type="email" 
+              name="user-email" 
+              id="user-email-form" 
+              class="modal-form-input" 
+              placeholder="exemple@gmail.com"
+              required />
+            </div>
+            <div class="modal-form-input-wrapper">
+              <label for="user-password-form" class="modal-form-label">Password</label>
+              </div>
+              <input 
+              type="password" 
+              name="user-password" 
+              id="user-password-form" 
+              class="modal-form-input" 
+              placeholder="use a strong password"
+                required />
+          </form>
+          <a class="lightbox-button-header btn-modal-close" data-action="close-lightbox-header">
+            <svg width="24" height="24">
+              <use href="./images/symbol-defs.svg#icon-close-2"></use>
+            </svg>
+          </a>
+        </div>
+      </div>
+    </div>
+
     <!-- modal footer -->
-    <div class="lightbox js-lightbox">
-      <div class="lightbox-overlay">
-        <div class="lightbox-content">
+    <div class="js-lightbox modal-lightbox">
+      <div class="lightbox-overlay overlay">
+        <div class="lightbox-content content-modal">
           <h2 class="title-header">Our team</h2>
           <ul class="footer-team-link"></ul>
-          <a class="lightbox-button" data-action="close-lightbox">
-            <svg width="24" height="24">
-              <use href="./images/symbol-defs.svg#icon-close"></use>
+          <a class="lightbox-button btn-modal-close" data-action="close-lightbox">
+            <svg width="48" height="48">
+              <use href="./images/symbol-defs.svg#icon-close-2"></use>
             </svg>
           </a>
         </div>
