+++ src/index.js
@@ -2,7 +2,7 @@ import './js/fetchAPIandMovieList/moviesListEventsHandler.js';
 import './scss/main.scss';
 import './js/pagination';
 import './js/modal';
-import './js/modal-footer';
+import './js/modal-minor';
 import './js/navigation';
 //import './js/lib';
 import './js/themes';
