--- src/scss/components/_modal-footer.scss
@@ -1,82 +1,0 @@
-:root {
-  --timing-function: cubic-bezier(0.4, 0, 0.2, 1);
-  --animation-duration: 250ms;
-}
-.lightbox {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  opacity: 0;
-  pointer-events: none;
-}
-.is-open {
-  overflow: hidden;
-  opacity: 1;
-  pointer-events: initial;
-}
-.lightbox-overlay {
-  z-index: 2;
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  background-color: rgba(0, 0, 0, 0.8);
-}
-.lightbox-content {
-  display: flex;
-
-  max-width: 100%;
-  max-height: 100%;
-
-  background-color: #fff;
-  transform: scale(0.95);
-  transition: transform var(--animation-duration) var(--timing-function);
-
-  @media screen and (max-width: 319px) {
-    margin-top: 20px;
-    margin-left: 20px;
-    margin-right: 20px;
-  }
-  @media screen and (min-width: 320px) {
-    margin-top: 231px;
-    margin-left: 75px;
-    margin-right: 75px;
-  }
-  @media screen and (min-width: 768px) {
-    margin-top: 100px;
-    margin-left: 71px;
-    margin-right: 71px;
-  }
-}
-.lightbox.is-open .lightbox-content {
-  transform: scale(1);
-}
-.lightbox-button {
-  z-index: 1111;
-  position: absolute;
-  top: 8px;
-  right: 20px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 48px;
-  height: 48px;
-  padding: 0;
-  margin: 0;
-  border: none;
-  border-radius: 50%;
-  cursor: pointer;
-  transition: background-color var(--animation-duration) var(--timing-function);
-  outline: none;
-  background-image: url('./../images/symbol-defs.svg#icon-close');
-  background-position: center;
-  background-size: 60%;
-  background-repeat: no-repeat;
-}
-.lightbox-button:hover,
-.lightbox-button:focus {
-  background-color: rgba(0, 0, 0, 0.5);
-}
