+++ src/scss/main.scss
@@ -12,8 +12,8 @@
 @import './components/input';
 @import './components/headerLib';
 
-@import './components/modal-footer';
 @import './components/modal';
+@import './components/modal-minor';
 @import './components/pagination';
 // @import './components/spinner';
